+++ lib/src/teledart/fetch/webhook.dart
@@ -30,13 +30,14 @@ class Webhook {
   io.SecurityContext _context;
 
   String url;
-  io.File certificate;
+  String secretPath;
+  int port;
   int max_connections;
   List<String> allowed_updates;
 
-  String secretPath;
-  int port;
+  io.File certificate;
   io.File privateKey;
+  bool uploadCertificate;
 
   StreamController<Update> _updateStreamController;
 
@@ -44,10 +45,10 @@ class Webhook {
   ///
   /// Throws [WebhookException] if [port] is not supported by Telegram
   /// or [max_connections] is less than 1 or greater than 100.
-  Webhook(this.telegram, this.url, this.secretPath,
-      {this.port = 443,
+  Webhook(this.telegram, this.url, this.secretPath, this.certificate,
       this.privateKey,
-      this.certificate,
+      {this.port = 443,
+      this.uploadCertificate = false,
       this.max_connections = 40,
       this.allowed_updates}) {
     if (![443, 80, 88, 8443].contains(this.port))
@@ -64,24 +65,19 @@ class Webhook {
       this.secretPath = '\/' + this.secretPath;
 
     // serup SecurityContext
-    if (privateKey != null && certificate != null) {
-      _context = new io.SecurityContext();
-      _context.usePrivateKeyBytes(privateKey.readAsBytesSync());
-      _context.useCertificateChainBytes(certificate.readAsBytesSync());
-    }
+    _context = new io.SecurityContext();
+    _context.useCertificateChainBytes(certificate.readAsBytesSync());
+    _context.usePrivateKeyBytes(privateKey.readAsBytesSync());
   }
 
   /// Set webhook on telegram server.
   Future<void> setWebhook() async {
-    // initialise server
-    Future<dynamic> serverFuture = _context == null
-        ? io.HttpServer.bind(io.InternetAddress.loopbackIPv4.address, port)
-        : io.HttpServer.bindSecure(
-            io.InternetAddress.loopbackIPv4.address, port, _context);
+    Future<dynamic> serverFuture = io.HttpServer.bindSecure(
+        io.InternetAddress.anyIPv4.address, port, _context);
 
     await serverFuture.then((server) => _server = server).then((_) {
       telegram.setWebhook('${this.url}:${this.port}${this.secretPath}',
-          certificate: certificate,
+          certificate: uploadCertificate ? certificate : null,
           max_connections: max_connections,
           allowed_updates: allowed_updates);
     }).catchError(
@@ -95,16 +91,18 @@ class Webhook {
           'Please use setWebhook() to initialise webhook before start webhook.');
     _server.listen((io.HttpRequest request) {
       if (request.method == 'POST' && request.uri.path == this.secretPath) {
-        request
-            .transform(utf8.decoder)
-            .join()
-            .then((data) => emitUpdate(new Update.fromJson(jsonDecode(data))));
-        request.response.write({'ok': true});
+        request.transform(utf8.decoder).join().then((data) {
+          emitUpdate(new Update.fromJson(jsonDecode(data)));
+          request.response
+            ..write(jsonEncode({'ok': true}))
+            ..close();
+        });
       } else {
-        request.response.statusCode = io.HttpStatus.methodNotAllowed;
-        request.response.write({'ok': false});
+        request.response
+          ..statusCode = io.HttpStatus.methodNotAllowed
+          ..write(jsonEncode({'ok': false}))
+          ..close();
       }
-      request.response.close();
     }).onError(
         (error) => new Future.error(new WebhookException(error.toString())));
   }